Add better Debian distribution compatibility. (#4)
Add better Debian distribution compatibility.

This required a lot of modifications to get Patroni to operate properly. The
Python 668 PEP essentially deprecates the "pip --user" installation method in
favor of venvs. So far only Debian appears to honor this, but there's no
guarantee of that being true in the future.

For now, the fix is to:

1. Install the Patroni binaries.
2. Create a venv for the requirements.
3. Install the requirements into this venv.
4. Add a wrapper for the patroni and patronictl commands which activates the
   environment prior to calling the "real" python script.

We could get around much of this if we just used the official Patroni package
rather than our own.

Other changes:

* Made core file debugging optional with new debug_pgedge parameter.
* Switch to some of the cleaner ansible variables rather than use ansible_facts.
* Ensure all host / inventory names are listed in /etc/hosts for managed systems. This is controlled by the new manage_host_file parameter.
* Moved the pgedge installer to the user's home dir to prevent a minor potential CVE.
* The non-HA sample playbook now only includes necessary tasks.