Extended package installation timeout to 5 minutes. (#8)
On freshly provisioned servers or VMs, background package management
may hold an extended lock that could disrupt an Ansible playbook which
needs to install packages. The default timeout of 30s is not sufficient
for this, and adding a wait step seems excessive. The lock_timeout
parameter is supported by both dnf and apt modules, and the package
module passes it along as is.

Addresses Jira issue EE-8.