Add multiple-database functionality.
If users want to have a cluster containing multiple databases, the
CLI tool won't do this natively. We need to add loops in strategic
areas to add all of the additional databases beyond the initial one
used to bootstrap the installation. This change hits all of the
required areas and also subsequently renames the `db_name` parmaeter
to `db_names`.

Additionally added a task to install cron on minimal systems.

Addresses Jira issue EE-12.