Removed fuser tasks and replaced them with standard retry loops.
This is a more "Ansible" solution which will retry the package
installation tasks until it is successful or the amount of retries
is exhausted. We may want to tweak the retry count or delay between
each retry, but this definitely works.

The issue is apparently that the Ansible package module doesn't even
try to wait for the lock to be acquired, it just aborts immediately.
It makes me wonder why they added the lock_timeout option in 2.12.

All package tasks now impose 5 retries with 20 seconds between each.