Fixing package management for Debian (#15)
Roles will retry package installation tasks until successful or the amount 
of retries is exhausted. We may want to tweak the retry count or delay
between each retry, but this definitely works.

The issue is apparently that the Ansible package module doesn't even
try to wait for the lock to be acquired, it just aborts immediately. If that's
the case, what does the lock_timeout option added in 2.12 even do?

All package tasks now impose 5 retries with 20 seconds between each.

---------

Signed-off-by: Paul Rothrock <paul.rothrock@pgedge.com>
Co-authored-by: Shaun Thomas <shaun@bonesmoses.org>