Revised to include the ability to add HAProxy backends from a dictionary.
Besides the basic HAProxy backend, it is now possible to add extra backends
based on the Patroni REST interface. New haproxy_extra_routes parameter
is a dictionary of all routes to add to the configuration file. For example:

```yaml
haproxy_extra_routes:
  replica:
    port: 5433
    lag: 1GB
  "read-only":
    port: 6543

Would create a backend to direct Postgres connections to port 5433 to any
registered replicas which are within 1GB of lag, and another backend on
port 6543 to replicas and the primary node depending on availability.