Switched etcd and Patroni to using upstream packages.
In the case of etcd, binary packages are provided directly on GitHub.
The new process downloads the tagged binary version, extracts, and
installs it. The download is also verified using the published checksum
on GitHub.

For Patroni, software is obtained from pypi.org via pipx. This ensures
all prerequisites are also included, and the required Python environment
always works as long as the binaries are invoked from the installation
path.

This addresses Jira issue EE-15.

Additionally, there is now a gratuitous restart of Postgres following
Patroni setup. This is to ensure all configuration changes applied to
the cluster are incorporated.

This addresses Jira issue EE-13.