Fix/ee 19/add redundant pipx environment file (#22)
Add redundant .bashrc.d environment file for pipx

While pipx was a huge upgrade for getting patroni installed and
operational, it doesn't specify which environment file it modifies,
and doesn't offer any parameters to target one specifically. Linux
distributions tend to differ on which will activate during an Ansible
shell, so the only way to be sure is to use our .bashrc.d/ environment
prepping directory like other roles in this collection.

That's a lot to say we just add .local/bin the the path in a file in
.bashrc.d/ under the (so far working) assumption that .bashrc will run
in all distributions.

Addresses Jira issue EE-19.