Nearly complete rewrite to switch from pgEdge CLI to native packages.
The CLI tool will eventually be deprecated, and pgEdge now produces
official OS distribution packages for RHEL and Debian family Linux
systems. This patch is an effort to switch to native integration with
these tools. This includes deferring to the postgres OS user when
possible, and no longer using the remote Ansible user as the owner of
database and other files.

Several new variables were introduced to account for default binary
and data directories, as well as operating system users. Existing
defaults have been demonstrated as functional on both RHEL 9 and Debian
12 cloud images.

Addresses JIRA Issue EE-22.